test(web): split community websocket suites
